backend/requirement.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
requests==2.32.3
pytest==8.2.2
alembic==1.13.2  # Optional for DB migrations
responses==0.25.3  # For mocking API calls in tests
mcp